# api/api-docs.yaml
openapi: 3.0.1
info:
  title: DL4DH Feeder APIs
  description: API endpoints
  version: 1.2.0
servers:
- url: https://feeder.nkp.cz
  description: Generated server url
paths:
  /api/search:
    post:
      tags:
      - search-api
      operationId: search
      parameters:
      - name: save
        in: query
        required: false
        schema:
          type: boolean
          default: false
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filter'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SearchDto'
  /api/search/hint:
    post:
      tags:
      - search-api
      operationId: hint
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
      - name: nameTagType
        in: query
        required: false
        schema:
          type: string
          enum:
          - NUMBERS_IN_ADDRESSES
          - GEOGRAPHICAL_NAMES
          - INSTITUTIONS
          - MEDIA_NAMES
          - NUMBER_EXPRESSIONS
          - ARTIFACT_NAMES
          - PERSONAL_NAMES
          - TIME_EXPRESSIONS
          - COMPLEX_PERSON_NAMES
          - COMPLEX_TIME_EXPRESSION
          - COMPLEX_ADDRESS_EXPRESSION
          - COMPLEX_BIBLIO_EXPRESSION
          - ALL
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
...
  /api/search/localPrintPDF:
    get:
      tags:
      - search-api
      operationId: print
      parameters:
      - name: pids
        in: query
        required: true
        schema:
          type: string
      - name: pagesize
        in: query
        required: false
        schema:
          type: string
      - name: imgop
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /api/search/history:
    get:
      tags:
      - search-api
      operationId: getAll
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          minimum: 0
          type: integer
          default: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          minimum: 1
          type: integer
          default: 10
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
          default:
          - "id,DESC"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageFilter'
...
  /api/item/{uuid}/children/search:
    get:
      tags:
      - item-api
      operationId: childrenSearch
      parameters:
      - name: uuid
        in: path
        required: true
        schema:
          type: string
      - name: q
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/ChildSearchDto'
  /api/item/{uuid}/children/search/hint:
    get:
      tags:
      - item-api
      operationId: childrenSearchHint
      parameters:
      - name: uuid
        in: path
        required: true
        schema:
          type: string
      - name: q
        in: query
        required: true
        schema:
          type: string
      - name: nameTagType
        in: query
        required: false
        schema:
          type: string
          enum:
          - NUMBERS_IN_ADDRESSES
          - GEOGRAPHICAL_NAMES
          - INSTITUTIONS
          - MEDIA_NAMES
          - NUMBER_EXPRESSIONS
          - ARTIFACT_NAMES
          - PERSONAL_NAMES
          - TIME_EXPRESSIONS
          - COMPLEX_PERSON_NAMES
          - COMPLEX_TIME_EXPRESSION
          - COMPLEX_ADDRESS_EXPRESSION
          - COMPLEX_BIBLIO_EXPRESSION
          - ALL
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                uniqueItems: true
                type: array
                items:
                  type: string
...
components:
  schemas:
    Filter:
      type: object
      properties:
        id:
          type: integer
          format: int64
        query:
          type: string
        availability:
          type: string
          enum:
          - PUBLIC
          - PRIVATE
          - ALL
        models:
          uniqueItems: true
          type: array
          items:
            type: string
            enum:
            - MONOGRAPH
            - PERIODICAL
            - MAP
            - GRAPHICS
            - ARCHIVAL
            - MANUSCRIPT
            - SHEETMUSIC
        keywords:
          uniqueItems: true
          type: array
          items:
            type: string
        authors:
          uniqueItems: true
          type: array
          items:
            type: string
        languages:
          uniqueItems: true
          type: array
          items:
            type: string
        collections:
          uniqueItems: true
          type: array
          items:
            type: string
        from:
          type: integer
          format: int32
        to:
          type: integer
          format: int32
        sort:
          type: string
          enum:
          - TITLE_ASC
          - CREATED_DESC
          - DATE_ASC
          - DATE_DESC
        enrichment:
          type: string
          enum:
          - ENRICHED
          - NOT_ENRICHED
          - ALL
        advancedFilterField:
          type: string
          enum:
          - TITLE
          - AUTHOR
          - KEYWORDS
          - NUMBERS_IN_ADDRESSES
          - GEOGRAPHICAL_NAMES
          - INSTITUTIONS
          - MEDIA_NAMES
          - NUMBER_EXPRESSIONS
          - ARTIFACT_NAMES
          - PERSONAL_NAMES
          - TIME_EXPRESSIONS
          - COMPLEX_PERSON_NAMES
          - COMPLEX_TIME_EXPRESSION
          - COMPLEX_ADDRESS_EXPRESSION
          - COMPLEX_BIBLIO_EXPRESSION
          - ALL_BASIC_METADATA
          - ALL_NAMETAG_DATA
          - NONE
        nameTagFilters:
          type: array
          items:
            $ref: '#/components/schemas/NameTagFilter'
        nameTagFacet:
          type: string
        start:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
    NameTagFilter:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum:
          - NUMBERS_IN_ADDRESSES
          - GEOGRAPHICAL_NAMES
          - INSTITUTIONS
          - MEDIA_NAMES
          - NUMBER_EXPRESSIONS
          - ARTIFACT_NAMES
          - PERSONAL_NAMES
          - TIME_EXPRESSIONS
          - COMPLEX_PERSON_NAMES
          - COMPLEX_TIME_EXPRESSION
          - COMPLEX_ADDRESS_EXPRESSION
          - COMPLEX_BIBLIO_EXPRESSION
          - ALL
        operator:
          type: string
          enum:
          - EQUAL
          - NOT_EQUAL
        values:
          uniqueItems: true
          type: array
          items:
            type: string
    PublicationDto:
      required:
      - authors
      - availability
      - date
      - enriched
      - model
      - pid
      - rootTitle
      - title
      type: object
      properties:
        model:
          type: string
        availability:
          type: string
        date:
          type: string
        authors:
          type: array
          items:
            type: string
        title:
          type: string
        pid:
          type: string
        rootTitle:
          type: string
        enriched:
          type: boolean
    PublicationsListDto:
      required:
      - docs
      - numFound
      - start
      type: object
      properties:
        numFound:
          type: integer
          format: int64
        start:
          type: integer
          format: int64
        docs:
          type: array
          items:
            $ref: '#/components/schemas/PublicationDto'
    SearchDto:
      type: object
      properties:
        documents:
          $ref: '#/components/schemas/PublicationsListDto'
        availableFilters:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: object
        availableNameTagFilters:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: object
...
    PageFilter:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Filter'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        pageable:
          $ref: '#/components/schemas/PageableObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/Sort'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
...
    ChildSearchDto:
      type: object
      properties:
        textOcr:
          type: array
          items:
            type: string
        nameTag:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...